Add files via upload
The dataset for this competition (both train and test) was generated from a deep learning model trained on the Stroke Prediction Dataset. Feature distributions are close to, but not exactly the same, as the original. Feel free to use the original dataset as part of this competition, both to explore differences as well as to see whether incorporating the original in training improves model performance.

Files train.csv - the training dataset; stroke is the binary target test.csv - the test dataset; your objective is to predict the probability of positive stroke sample_submission.csv - a sample submission file in the correct format